+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fix
+
+- [TD-6950] Delete only the specified domains and roles
+
 ## [6.15.0] 2024-11-13
 
 ### Added